added deactivation of trades restendpoints via adminsettings

+++ src/main/java/de/hsrm/mi/swtp/exchangeplatform/controller/TradeOffersRestController.java
@@ -6,6 +6,7 @@ import de.hsrm.mi.swtp.exchangeplatform.model.data.TradeOffer;
 import de.hsrm.mi.swtp.exchangeplatform.model.rest.TradeRequest;
 import de.hsrm.mi.swtp.exchangeplatform.service.rest.StudentService;
 import de.hsrm.mi.swtp.exchangeplatform.service.rest.TradeOfferService;
+import de.hsrm.mi.swtp.exchangeplatform.service.settings.AdminSettingsService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
@@ -30,6 +31,7 @@ public class TradeOffersRestController {
 	
 	TradeOfferService tradeOfferService;
 	StudentService studentService;
+	AdminSettingsService adminSettingsService;
 	
 	/**
 	 * DELETE request handler.
@@ -49,18 +51,21 @@ public class TradeOffersRestController {
 	public ResponseEntity deleteTradeOffer(@ApiParam(value = "Numeric ID of the student", required = true) @PathVariable("studentId") long studentId,
 										   @ApiParam(value = "Numeric ID of the tradeoffer", required = true) @PathVariable("tradeId") long tradeId
 										  ) {
-		log.info(String.format("DELETE Request Student: %d TradeOffer: %d", studentId, tradeId));
-		try {
-			if(tradeOfferService.deleteTradeOffer(studentId, tradeId)) {
-				log.info(String.format("DELETE Request successful Student: %d TradeOffer: %d", studentId, tradeId));
-				return new ResponseEntity<>(HttpStatus.OK);
+		if(adminSettingsService.isTradesActive()) {
+			log.info(String.format("DELETE Request Student: %d TradeOffer: %d", studentId, tradeId));
+			try {
+				if(tradeOfferService.deleteTradeOffer(studentId, tradeId)) {
+					log.info(String.format("DELETE Request successful Student: %d TradeOffer: %d", studentId, tradeId));
+					return new ResponseEntity<>(HttpStatus.OK);
+				}
+			} catch(TradeOfferNotFoundException ex) {
+				log.info(String.format("ERROR while DELETE Request Student: %d TradeOffer: %d - Entity not found", studentId, tradeId));
+				return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 			}
-		} catch(TradeOfferNotFoundException ex) {
-			log.info(String.format("ERROR while DELETE Request Student: %d TradeOffer: %d - Entity not found", studentId, tradeId));
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-		}
-		log.info(String.format("ERROR while DELETE Request Student: %d TradeOffer: %d - Student isn't owner of entity", studentId, tradeId));
-		return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+			log.info(String.format("ERROR while DELETE Request Student: %d TradeOffer: %d - Student isn't owner of entity", studentId, tradeId));
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		} else return new ResponseEntity(HttpStatus.SERVICE_UNAVAILABLE);
+		
 	}
 	
 	/**
@@ -81,21 +86,23 @@ public class TradeOffersRestController {
 			@ApiParam(value = "Object containing ID's of student, wanted and offered timeslot", required = true) @Valid TradeRequest tradeRequest,
 			BindingResult bindingResult
 													  ) {
-		
-		if(bindingResult.hasErrors()) {
-			log.info("Malformed traderequest");
-			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-		}
-		log.info(String.format("POST Request create new traderequest: Requester: %d Offered: %d, Seek: %d",
-							   tradeRequest.getOfferedByStudentMatriculationNumber(), tradeRequest.getOfferedTimeslotId(), tradeRequest.getWantedTimeslotId()
-							  ));
-		var persistedTradeOffer = tradeOfferService.createTradeOffer(tradeRequest.getOfferedByStudentMatriculationNumber(), tradeRequest.getOfferedTimeslotId(),
-																	 tradeRequest.getWantedTimeslotId()
-																	);
-		log.info(String.format("POST Request successful: created new tradeoffer with id: %d requester: %d", persistedTradeOffer.getId(),
-							   persistedTradeOffer.getOfferer().getStudentId()
-							  ));
-		return new ResponseEntity<>(persistedTradeOffer, HttpStatus.OK);
+		if(adminSettingsService.isTradesActive()) {
+			if(bindingResult.hasErrors()) {
+				log.info("Malformed traderequest");
+				return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+			}
+			log.info(String.format("POST Request create new traderequest: Requester: %d Offered: %d, Seek: %d",
+								   tradeRequest.getOfferedByStudentMatriculationNumber(), tradeRequest.getOfferedTimeslotId(),
+								   tradeRequest.getWantedTimeslotId()
+								  ));
+			var persistedTradeOffer = tradeOfferService.createTradeOffer(tradeRequest.getOfferedByStudentMatriculationNumber(),
+																		 tradeRequest.getOfferedTimeslotId(), tradeRequest.getWantedTimeslotId()
+																		);
+			log.info(String.format("POST Request successful: created new tradeoffer with id: %d requester: %d", persistedTradeOffer.getId(),
+								   persistedTradeOffer.getOfferer().getStudentId()
+								  ));
+			return new ResponseEntity<>(persistedTradeOffer, HttpStatus.OK);
+		} else return new ResponseEntity<>(HttpStatus.SERVICE_UNAVAILABLE);
 	}
 	
 	@PostMapping
@@ -104,15 +111,17 @@ public class TradeOffersRestController {
 							@ApiResponse(code = 403, message = "unauthorized trade attempt"),
 							@ApiResponse(code = 400, message = "malformed trade request") })
 	public ResponseEntity<TimeTable> requestTrade(@Valid TradeRequest tradeRequest) {
-		log.info(String.format("Traderequest of student: %d for timeslot: %d, offer: %d", tradeRequest.getOfferedByStudentMatriculationNumber(),
-							   tradeRequest.getOfferedTimeslotId(), tradeRequest.getWantedTimeslotId()
-							  ));
-		var timeslot = tradeOfferService.tradeTimeslots(tradeRequest.getOfferedByStudentMatriculationNumber(), tradeRequest.getOfferedTimeslotId(),
-														tradeRequest.getWantedTimeslotId()
-													   );
-		TimeTable timetable = new TimeTable();
-		timetable.setId(tradeRequest.getOfferedByStudentMatriculationNumber());
-		timetable.setTimeslots(studentService.getById(tradeRequest.getOfferedByStudentMatriculationNumber()).getTimeslots());
-		return new ResponseEntity<>(timetable, HttpStatus.OK);
+		if(adminSettingsService.isTradesActive()) {
+			log.info(String.format("Traderequest of student: %d for timeslot: %d, offer: %d", tradeRequest.getOfferedByStudentMatriculationNumber(),
+								   tradeRequest.getOfferedTimeslotId(), tradeRequest.getWantedTimeslotId()
+								  ));
+			var timeslot = tradeOfferService.tradeTimeslots(tradeRequest.getOfferedByStudentMatriculationNumber(), tradeRequest.getOfferedTimeslotId(),
+															tradeRequest.getWantedTimeslotId()
+														   );
+			TimeTable timetable = new TimeTable();
+			timetable.setId(tradeRequest.getOfferedByStudentMatriculationNumber());
+			timetable.setTimeslots(studentService.getById(tradeRequest.getOfferedByStudentMatriculationNumber()).getTimeslots());
+			return new ResponseEntity<>(timetable, HttpStatus.OK);
+		} else return new ResponseEntity<>(HttpStatus.SERVICE_UNAVAILABLE);
 	}
 }
